pyproject.toml
[project]
name = "communicator-bridge"
version = "0.1.0"
description = "Sistema de integração de canais de comunicação"
authors = [
    {name = "Lucas de Freitas",email = "lfreitas.dev@pm.me"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "django (>=5.2.5,<6.0.0)",
    "djangorestframework (>=3.16.1,<4.0.0)",
    "django-cors-headers (>=4.7.0,<5.0.0)",
    "djangorestframework-simplejwt (>=5.5.1,<6.0.0)",
    "django-auditlog (>=3.2.1,<4.0.0)",
    "requests (>=2.32.4,<3.0.0)",
    "psycopg2-binary (>=2.9.9,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
black = "^25.1.0"
isort = "^6.0.1"

[tool.black]
line-length = 88
target-version = ['py310']

[tool.isort]
profile = "black"
line_length = 88
known_first_party = ["core"]

